backend/requirements/base.txt
# Core Django
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# JWT Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.11.0

# API schema / docs
drf-spectacular==0.27.1

# Nested routers
drf-nested-routers==0.94.1

# Static files
whitenoise==6.6.0

# Async tasks
celery==5.3.6
redis==5.3.1

# AI
openai==1.12.0

# Email (SendGrid)
sendgrid==6.11.0

# PDF generation (evidence export)
reportlab==4.1.0

# Content extraction
pypdf==4.2.0
python-pptx==1.0.2
Pillow==10.3.0
pytesseract==0.3.10

# Web search + scraping
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0

# Environment
python-dotenv==1.0.1

# WSGI utils
asgiref==3.7.2
sqlparse==0.4.4
